refactor(component-linter): Consolidate 10 RunView/RunQuery rules into 3
Replace 10 overlapping rules with 3 focused, non-overlapping rules:
- runview-call-validation: All RunView/RunViews call-site validation
  (property names/types, EntityName existence, argument structure)
- runquery-call-validation: All RunQuery call-site validation
  (property names/types, Parameters format/names/types/nulls,
  CategoryPath, SQL injection, variable resolution via TypeContext)
- data-result-validation: All RunView + RunQuery result usage validation
  (invalid property access, spread, array methods, Success guards,
  setState patterns, destructuring)

Eliminates duplicated "find call → extract config → handle variable"
boilerplate that caused the false positive bug to need fixing in
multiple places. Includes 18-test consolidation audit verifying each
specific check produces the expected violation.
