+++ packages/React/test-harness/src/lib/runtime-rules/index.ts
@@ -10,6 +10,7 @@ import './component-not-in-dependencies';
 import './component-props-validation';
 import './component-usage-without-destructuring';
 import './datagrid-field-validation';
+import './data-result-validation';
 import './dependency-prop-validation';
 import './dependency-shadowing';
 import './entity-field-access-validation';
@@ -33,20 +34,12 @@ import './prefer-jsx-syntax';
 import './prop-state-sync';
 import './property-name-consistency';
 import './query-field-validation';
-import './query-param-null-check';
-import './query-parameter-type-validation';
 import './query-result-field-access-validation';
 import './react-component-naming';
 import './react-hooks-rules';
 import './required-queries-not-called';
-import './runquery-missing-categorypath';
-import './runquery-parameters-validation';
-import './runquery-runview-spread-operator';
-import './runquery-runview-validation';
-import './runview-entity-validation';
-import './runview-result-null-safety';
-import './runview-runquery-result-validation';
-import './runview-runquery-valid-properties';
+import './runquery-call-validation';
+import './runview-call-validation';
 import './saved-user-settings-pattern';
 import './server-reload-on-client-operation';
 import './single-function-only';
